Move `FolderType`, `Folder` and `FolderDetails` to `:core:mail:folder:api` module
