+++ core/mail/folder/api/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    id(ThunderbirdPlugins.Library.jvm)
+    alias(libs.plugins.android.lint)
+}
+
+dependencies {
+    implementation(projects.mail.common)
+
+    testImplementation(projects.core.testing)
+}
