+++ core/mail/folder/api/src/main/kotlin/app/k9mail/core/mail/folder/api/Folder.kt
@@ -1,4 +1,4 @@
-package app.k9mail.legacy.folder
+package app.k9mail.core.mail.folder.api
 
 data class Folder(
     val id: Long,
