+++ legacy/core/src/main/java/com/fsck/k9/mailstore/SpecialFolderUpdater.kt
@@ -1,9 +1,9 @@
 package com.fsck.k9.mailstore
 
 import app.k9mail.core.common.mail.Protocols
+import app.k9mail.core.mail.folder.api.FolderType
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.account.Account.SpecialFolderSelection
-import app.k9mail.legacy.folder.FolderType
 import app.k9mail.legacy.folder.RemoteFolder
 import app.k9mail.legacy.mailstore.FolderRepository
 import com.fsck.k9.Preferences
