+++ legacy/folder/src/main/kotlin/app/k9mail/legacy/folder/RemoteFolder.kt
@@ -1,5 +1,7 @@
 package app.k9mail.legacy.folder
 
+import app.k9mail.core.mail.folder.api.FolderType
+
 data class RemoteFolder(
     val id: Long,
     val serverId: String,
