+++ legacy/ui/folder/build.gradle.kts
@@ -9,6 +9,8 @@ android {
 dependencies {
     implementation(projects.core.ui.legacy.designsystem)
 
+    implementation(projects.core.mail.folder.api)
+
     implementation(projects.legacy.account)
     implementation(projects.legacy.folder)
     implementation(projects.legacy.mailstore)
