+++ legacy/ui/folder/src/main/java/app/k9mail/legacy/ui/folder/FolderIconProvider.kt
@@ -1,7 +1,7 @@
 package app.k9mail.legacy.ui.folder
 
+import app.k9mail.core.mail.folder.api.FolderType
 import app.k9mail.core.ui.legacy.designsystem.atom.icon.Icons
-import app.k9mail.legacy.folder.FolderType
 
 class FolderIconProvider {
     fun getFolderIcon(type: FolderType): Int = when (type) {
