+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/managefolders/FolderSettingsDataStore.kt
@@ -1,8 +1,8 @@
 package com.fsck.k9.ui.managefolders
 
 import androidx.preference.PreferenceDataStore
+import app.k9mail.core.mail.folder.api.FolderDetails
 import app.k9mail.legacy.account.Account
-import app.k9mail.legacy.folder.FolderDetails
 import app.k9mail.legacy.mailstore.FolderRepository
 import com.fsck.k9.mail.FolderClass
 import kotlinx.coroutines.CoroutineScope
