+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/messagedetails/MessageDetailsUi.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.ui.messagedetails
 
 import android.net.Uri
-import app.k9mail.legacy.folder.FolderType
+import app.k9mail.core.mail.folder.api.FolderType
 import com.fsck.k9.mail.Address
 import com.fsck.k9.view.MessageCryptoDisplayStatus
