+++ settings.gradle.kts
@@ -96,6 +96,10 @@ include(
 )
 
 include(
+    ":core:mail:folder:api",
+)
+
+include(
     ":legacy:account",
     ":legacy:common",
     ":legacy:core",
